attach completely created

+++ src/main/java/com/example/repositiry/AttachRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AttachRepository extends CrudRepository<AttachEntity, Integer> {
+public interface AttachRepository extends CrudRepository<AttachEntity, Long> {
 
 }
